+++ spec/per_session_config_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe Capybara::SessionConfig do
       session = Capybara::Session.new(:rack_test, TestApp)
       [:default_host, :app_host, :save_and_open_page_path,
        :always_include_port, :run_server, :default_selector, :default_max_wait_time, :ignore_hidden_elements,
-       :automatic_reload, :match, :exact, :raise_server_errors, :visible_text_only, :wait_on_first_by_default,
+       :automatic_reload, :match, :exact, :raise_server_errors, :visible_text_only,
        :automatic_label_click, :enable_aria_label,
        :save_path, :exact_options, :asset_host].each do |m|
          expect(session.config.public_send(m)).to eq Capybara.public_send(m)