+++ History.md
@@ -1,5 +1,13 @@
 # Version 3.0.0
+
 Release date: unreleased
+
+### Added
+
+* Ruby 2.2+ required
+* RSpec 2.x no longer supported
+* Deprecated methods removed
+* `first` now raises ElementNotFound by default instead of returning nil, when no matches are found  - Issue #1507
 
 # Version 2.16.0
 Release date: 2017-11-13
@@ -223,6 +231,7 @@ Release date: 2016-10-05
 Release date: 2016-09-29
 
 ### Fixed
+
 * :label built-in selector finds nested label/control by control id if the label has no 'for' attribute [Thomas Walpole]
 * Warning issued if an unknown selector type is specified [Thomas Walpole]